Move PlateTabBar connect
