+++ ui/src/components/editor/PlateTabBar/index.js
@@ -1,2 +1,1 @@
-import { PlateTabBar } from './connect';
-export { PlateTabBar };
+export { PlateTabBar } from './PlateTabBar';
