+++ ui/src/components/index.js
@@ -4,3 +4,4 @@ export { ActivityHeader } from './activity/ActivityHeader';
 export { PlateSizeForm } from './activity/PlateSizeForm';
 export { Print } from './print';
 export { Editor } from './editor/Editor';
+export { PlateTabBar } from './editor/PlateTabBar';
