+++ ui/src/implementations/kapture-backend/components/index.js
@@ -5,3 +5,4 @@ export { ActivityDetails } from './ActivityDetails';
 export { PlateSizeForm } from '../../../components';
 export { Print } from './PrintConnected';
 export { Editor } from './EditorConnected';
+export { PlateTabBar } from './PlateTabBarConnected';
